+++ pages/folder/index.tsx
@@ -5,7 +5,6 @@ import Header from "@/components/Header/Header";
 import LinkCreator from "@/components/LinkCreator/LinkCreator";
 import Contents from "@/components/Contents/Contents";
 import CardSearchBar from "@/components/Contents/CardSearchBar/CardSearchBar";
-import FolderCollection from "@/components/Contents/FolderCollection/FolderCollection";
 import CardList from "@/components/Contents/CardList/CardList";
 import Footer from "@/components/Footer/Footer";
 import Modal from "@/components/Modal/Modal";
@@ -16,7 +15,7 @@ import {
 } from "@/hooks/Folder.hook";
 import { useSearchBar } from "@/hooks/useSearchBar";
 import { getAccessToken } from "@/utils/getAccessToken";
-import { getRefinedFolderList } from "@/apis/services";
+import FolderList from "@/components/Contents/FolderList/FolderList";
 
 export default function Folder() {
     const router = useRouter();
@@ -56,7 +55,7 @@ export default function Folder() {
                     onInputChange={handleInputChange}
                     resetInputValue={resetInputValue}
                 />
-                <FolderCollection
+                <FolderList
                     onButtonClick={showModal}
                     onOverviewFolderButtonClick={
                         handleOverviewFolderButtonClick
