+++ pages/signup/index.tsx
@@ -1,6 +1,7 @@
 import { Form } from "@/components/Sign/Form";
-import { Header } from "@/components/Sign/Header";
-import { SnsLogin } from "@/components/Sign/SnsLogin";
+import { Header } from "@/components/Sign/Shared/Header";
+import { SnsLogin } from "@/components/Sign/Shared/SnsLogin";
+import { SignUpForm } from "@/components/Sign/SignUp/Form";
 import { useRouter } from "next/router";
 import { useEffect } from "react";
 import styled from "styled-components";
@@ -17,7 +18,7 @@ export default function Signup() {
     return (
         <Wrapper>
             <Header currentPath="signup" />
-            <Form currentPath="signup" />
+            <SignUpForm />
             <SnsLogin />
         </Wrapper>
     );
